AnimalUpdates Notification ViewController
Added viewcontroller to show custom notifications for animal updates.
There is an optional image that can be added. If the image is blank the
view will resize. The view also expands with the amount of text in the
notification.

Changed tableView to only animate once initially. I’ve been back and
forth on this, but I changed this so that updates to the database
wouldn’t reload the whole collectionView.