Bug Fixes for v2.0.2
Added social media buttons to link to the aquarium webpages.
Added extension for openUrl to loop through and check if the app is
installed, and if not, open webpage in Safari.

iPhone 5 constraint fixes:
Fixed constraints with home screen
Fixed constraints with QR CollectionView
Fixed constraints with contact us view controller
Fixed constraints for Add A Membership viewcontroller